Dead Pokemon persist on the board ...but are ignored for movement and combat.
Need to persist to allow death animations to happen.

To allow for dead Pokemon to persist, multiple pieces are allowed on one tile now.
Although, only one should be alive.

Have added an "id" to a Pokémon piece, to be used as a key when iterating over pieces on a tile.